+++ stag_lib/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - [Issue #70](https://github.com/staglibrary/stag/issues/70): Add general stochastic block model method
+- [Issue #69](https://github.com/staglibrary/stag/issues/69): Add methods to get the ground truth labels for SBM graphs
 
 ## [1.0.0] - 2023-02-16
 ### Changed